+++ src/app/services/selector_matcher.js
@@ -186,9 +186,11 @@ function getNodesByTestType(elements, test_type) {
 
         if (node.resource_type != 'test') {
             return false;
-        } else if (_.includes(node.tags, 'schema') && test_type == 'schema') {
+        // generic tests have `test_metadata`, singular tests do not
+        // for backwards compatibility, keep supporting old test_type names
+        } else if node.hasOwnProperty('test_metadata') && test_type in ('schema', 'generic') {
             nodes.push(node);
-        } else if (_.includes(node.tags, 'data') && test_type == 'data') {
+        } else if !node.hasOwnProperty('test_metadata') && test_type in ('data', 'singular') {
             nodes.push(node);
         }
     });